leitura fix, ranreal testado! minor fix int->float

+++ Edge/Edge.cpp
@@ -4,7 +4,7 @@
 
 //Edge::Edge(Node *nodeHead, Node *tailNode, int weight, Graph *graph) {
 
-Edge::Edge(Node *headNode, Node *tailNode, int weight, Graph *graph) {
+Edge::Edge(Node *headNode, Node *tailNode, float weight, Graph *graph) {
     this->headNode = headNode;
     this->tailNode = tailNode;
     this->weight = weight;
